--- Seven Bridges/Views/GraphView.swift
@@ -1,48 +1,0 @@
-import UIKit
-
-public class GraphView: UIView {
-    
-    weak var parentVC: ViewController?
-    var graph = Graph()
-    
-    func clear() {
-        graph = Graph()
-        updatePropertiesToolbar()
-    }
-    
-    private func updatePropertiesToolbar() {
-        // hide the toolbar if no nodes are selected
-        if graph.selectedNodes.isEmpty {
-            parentVC?.trashButton.isEnabled = false
-        } else {
-            parentVC?.trashButton.isEnabled = true
-        }
-        
-        // detect a selected edge between two nodes
-        // if nil, disable UI elements related to a selected edge
-        if let edge = graph.selectedEdge {
-            parentVC?.edgeWeightIndicator.title = String(edge.weight)
-            
-            parentVC?.edgeWeightMinusButton.title = "-"
-            parentVC?.edgeWeightMinusButton.isEnabled = true
-            
-            parentVC?.edgeWeightPlusButton.title = "+"
-            parentVC?.edgeWeightPlusButton.isEnabled = true
-            
-            parentVC?.removeEdgeButton.title = "Remove \(edge)"
-            parentVC?.removeEdgeButton.isEnabled = true
-        } else {
-            parentVC?.edgeWeightIndicator.title = ""
-            parentVC?.edgeWeightIndicator.isEnabled = false
-            
-            parentVC?.edgeWeightMinusButton.title = ""
-            parentVC?.edgeWeightMinusButton.isEnabled = false
-            
-            parentVC?.edgeWeightPlusButton.title = ""
-            parentVC?.edgeWeightPlusButton.isEnabled = false
-            
-            parentVC?.removeEdgeButton.title = ""
-            parentVC?.removeEdgeButton.isEnabled = false
-        }
-    }
-}
