--- Seven Bridges/Views/SelectingGraphView.swift
@@ -1,3 +1,0 @@
-import UIKit
-
-public class SelectingGraphView: GraphView {}
